# StudentPortal.API/bruno/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "StudentPortal API",
    "description": "PRN232 Lab 1 – RESTful API for managing Students, Courses, Subjects and Semesters.",
    "contact": {
      "name": "PRN232 Student",
      "email": "student@fpt.edu.vn"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get all students with pagination, search, sorting, select, and expand options.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortdescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "$select",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "$expand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a paged list of students.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectPagedResponseApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Create a new student.",
        "requestBody": {
          "description": "Student data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or email already registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get a single student by ID (includes enrolled courses).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Student ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "summary": "Update an existing student.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Student ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or email conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "Delete a student by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Student ID.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStudentRequest": {
        "required": [
          "email",
          "fullName"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string",
            "example": "Nguyen Van An"
          },
          "email": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "format": "email",
            "example": "an.nguyen@fpt.edu.vn"
          }
        },
        "additionalProperties": false,
        "description": "Payload for creating a new Student."
      },
...
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
      "ObjectPagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Pagination metadata wrapper."
      },
      "ObjectPagedResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ObjectPagedResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
      "StudentResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Base student response — no navigation properties."
      },
      "StudentResponseApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/StudentResponse"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard API response envelope."
      },
...
      "UpdateStudentRequest": {
        "required": [
          "email",
          "fullName"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string",
            "example": "Nguyen Van An"
          },
          "email": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "format": "email",
            "example": "an.nguyen@fpt.edu.vn"
          }
        },
        "additionalProperties": false,
        "description": "Payload for updating an existing Student."
      },
...
    }
  }
}